terraform/modules/data/function_limpieza/requirements.txt
pandas
google-cloud-bigquery
flask
pyarrow
db-dtypes